Henrik

+++ tools/create_db.py
@@ -33,11 +33,11 @@ from definitions import midvatten_defs as defs
 
 class newdb():
 
-    def __init__(self, verno, user_select_CRS='y', EPSG_code=u'4326'):
+    def __init__(self, verno, user_select_CRS='y', EPSG_code=u'4326', delete_srids=True):
         self.dbpath = ''
-        self.create_new_db(verno,user_select_CRS,EPSG_code)  #CreateNewDB(verno)
+        self.create_new_db(verno,user_select_CRS,EPSG_code, delete_srids)  #CreateNewDB(verno)
 
-    def create_new_db(self, verno, user_select_CRS='y', EPSG_code=u'4326'):  #CreateNewDB(self, verno):
+    def create_new_db(self, verno, user_select_CRS='y', EPSG_code=u'4326', delete_srids=True):  #CreateNewDB(self, verno):
         """Open a new DataBase (create an empty one if file doesn't exists) and set as default DB"""
         set_locale = self.ask_for_locale()
         if set_locale == u'cancel':
@@ -119,15 +119,16 @@ class newdb():
                         utils.MessagebarAndLog.critical("sqlite error, see qgis Log Message Panel", 'Failed to create DB! sql failed: \n%serror msg: %s\n\n'%(line ,str(e)), duration=5)
 
                 #utils.MessagebarAndLog.info(bar_msg=u"epsgid: " + utils.returnunicode(EPSGID))
-                try:#spatial_ref_sys_aux not implemented until spatialite 4.3
-                    self.cur.execute(r"""delete from spatial_ref_sys_aux where srid NOT IN ('%s', '4326')""" % EPSGID)
-                except:
-                    pass
-                delete_srid_sql = r"""delete from spatial_ref_sys where srid NOT IN ('%s', '4326')""" % EPSGID
-                try:
-                    self.cur.execute(delete_srid_sql)
-                except:
-                    utils.MessagebarAndLog.info(log_msg=u'Removing srids failed using: ' + str(delete_srid_sql))
+                if delete_srids:
+                    try:#spatial_ref_sys_aux not implemented until spatialite 4.3
+                        self.cur.execute(r"""delete from spatial_ref_sys_aux where srid NOT IN ('%s', '4326')""" % EPSGID)
+                    except:
+                        pass
+                    delete_srid_sql = r"""delete from spatial_ref_sys where srid NOT IN ('%s', '4326')""" % EPSGID
+                    try:
+                        self.cur.execute(delete_srid_sql)
+                    except:
+                        utils.MessagebarAndLog.info(log_msg=u'Removing srids failed using: ' + str(delete_srid_sql))
 
                 self.insert_datadomains(set_locale)
